Move intime logic to golib
