+++ README.md
@@ -41,4 +41,3 @@ manually import the package in the import section by adding `"github.com/wowmate
 #### upload infrastructure
 
 ![upload infrastructure](misc/etl.png)
-
